Fixed reflections with cubemap
Remaking the cubemap asset fixed issues. The initial issue with the seams of the cubemap not lining up was fixed by swapping the positive and negative x images 'px' and 'nx' with each other when creating the legacy cubemap asset.

Note: When creating a legacy cubemap asset, make sure to set the setting you want for it before importing the images. It can cause issues when you change settings after already adding the images to the asset. This also goes for a lot of the other steps, so when in doubt, restart the process of setting up the skybox and try again without making too many changes to steps you already completed.